Seção 9: DTOs e Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -24,7 +24,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var eventos = await _eventoService.GetAllEventosAsync(true);
-                 if (eventos == null) return NotFound("Nenhum evento encontrado.");
+                 if (eventos == null) return NoContent();
 
                  return Ok(eventos);
             }
@@ -41,7 +41,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                 if (evento == null) return NotFound("Evento por Id não encontrado.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -58,7 +58,7 @@ namespace ProEventos.API.Controllers
             try
             {
                  var eventos = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                 if (eventos == null) return NotFound("Eventos por tema não encontrados.");
+                 if (eventos == null) return NoContent();
 
                  return Ok(eventos);
             }
@@ -70,12 +70,12 @@ namespace ProEventos.API.Controllers
         }        
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                  var evento = await _eventoService.AddEventos(model);
-                 if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -87,12 +87,12 @@ namespace ProEventos.API.Controllers
         }
         
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                  var evento = await _eventoService.UpdateEvento(id, model);
-                 if (evento == null) return BadRequest("Erro ao tentar atualizar evento.");
+                 if (evento == null) return NoContent();
 
                  return Ok(evento);
             }
@@ -108,9 +108,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                         Ok("Deletado") : 
-                        BadRequest("Evento não deletado");
+                        throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento.");
             }
             catch (Exception ex)
             {
